+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `stickyHeaderTabController(_, didUpdateCompoundHeaderHeight:)` delegate method.
+
+### Fixed
+- Content jump due to short tab content being pinned to tab bar after switching tabs.
 
 ## [0.1.0] - 2017-10-05
 ### Added